Task_7.4
Took 58 minutes

+++ kodilla-stream/src/main/java/com/kodilla/stream/sand/SandStorage.java
@@ -1,6 +1,8 @@
 package com.kodilla.stream.sand;
 
+import java.math.BigDecimal;
+
 public interface SandStorage {
 
-    long getSandBeansQuantity();
+    BigDecimal getSandBeansQuantity();
 }
